chore(examples): regen bundled codegen for proptest_gen + adapter fixes
The prior six commits change codegen output for proptest harnesses (and
two adapter helpers used by the kani / guards backends), so the bundled
example artifacts drift against the new emission. Catches up the six
affected files so `qedgen check --regen-drift` stays clean.

Per-file rationale:

  examples/rust/lending/tests/proptest.rs        — adds `status:` line
                                                   to lifecycle init
                                                   seed-state (commit 3)

  examples/rust/multisig/tests/proptest.rs       — type-aware seed-state
  examples/rust/multisig/programs/tests/proptest.rs init for the
                                                   `voted: Map[32] U8`
                                                   field (commit 3) +
                                                   `status:` line

  examples/rust/percolator/tests/proptest.rs    — `accounts: Map[1024]
                                                   Account` gets the
                                                   record-aware default
                                                   `[Account::default();
                                                   1024]` (commit 3) +
                                                   `status: Status::
                                                   Active` (initial
                                                   lifecycle state, not
                                                   hardcoded
                                                   `Uninitialized`,
                                                   commit 3) + symmetric
                                                   i128 widening on
                                                   `pnl + s.I` add
                                                   guards (commit 1)

  examples/rust/percolator/tests/kani.rs        — symmetric i128
                                                   widening on the same
                                                   `pnl` add-guard
                                                   pattern (commit 1)

  examples/rust/percolator/programs/src/guards.rs — symmetric i128
                                                   widening on the
                                                   `(a + b) >= 0`
                                                   underwater check that
                                                   appears in `add_user`,
                                                   `add_lp`, `withdraw`,
                                                   and the three
                                                   `liquidate_*` arms
                                                   (commit 1)

(The escrow example is no longer affected — commit 4's gating means
mul_div helpers are only inlined when the spec uses them, which escrow
doesn't, and commit 6's gating means recursion_limit is only emitted for
specs with >32 fields, which escrow has 3.)

Drift detection: `qedgen check --regen-drift` is clean for these six
files; broader bundled-example drift unrelated to this branch is left
in place to keep the diff focused.

+++ examples/rust/percolator/tests/kani.rs
@@ -76,7 +76,7 @@ fn account_solvent(_s: &State) -> bool {
 // ============================================================================
 
 fn add_user(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -88,7 +88,7 @@ fn add_user(s: &mut State, i: usize) -> bool {
 }
 
 fn add_lp(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -148,7 +148,7 @@ fn deposit(s: &mut State, i: usize, amount: u128) -> bool {
 }
 
 fn withdraw(s: &mut State, i: usize, amount: u128) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((amount) as i128))) {
+    if !((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((amount) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -239,7 +239,7 @@ fn execute_trade(s: &mut State, a: usize, b: usize, size_q: i128, exec_price: u6
 }
 
 fn liquidate_case_0(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128)) && (false)) {
+    if !((s.accounts[(i) as usize].active == 1) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128)) && (false)) {
         return false;
     }
     if s.status != Status::Active {
@@ -250,7 +250,7 @@ fn liquidate_case_0(s: &mut State, i: usize) -> bool {
 }
 
 fn liquidate_case_1(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl + ((s.I) as i128) >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -262,7 +262,7 @@ fn liquidate_case_1(s: &mut State, i: usize) -> bool {
 }
 
 fn liquidate_otherwise(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl + ((s.I) as i128) >= ((0) as i128))) && (false)) {
+    if !((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (!(((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) && (false)) {
         return false;
     }
     if s.status != Status::Active {
@@ -324,7 +324,7 @@ fn verify_add_user_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))));
     assert!(!add_user(&mut s, i),
         "add_user must reject when guard is violated");
 }
@@ -342,7 +342,7 @@ fn verify_add_lp_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))));
     assert!(!add_lp(&mut s, i),
         "add_lp must reject when guard is violated");
 }
@@ -416,7 +416,7 @@ fn verify_withdraw_rejects_invalid() {
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
     let amount: u128 = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((amount) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((amount) as i128))));
     assert!(!withdraw(&mut s, i, amount),
         "withdraw must reject when guard is violated");
 }
@@ -511,7 +511,7 @@ fn verify_liquidate_case_0_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128)) && (false)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128)) && (false)));
     assert!(!liquidate_case_0(&mut s, i),
         "liquidate_case_0 must reject when guard is violated");
 }
@@ -529,7 +529,7 @@ fn verify_liquidate_case_1_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) && (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl + ((s.I) as i128) >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))));
     assert!(!liquidate_case_1(&mut s, i),
         "liquidate_case_1 must reject when guard is violated");
 }
@@ -547,7 +547,7 @@ fn verify_liquidate_otherwise_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))) && (!(((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl + ((s.I) as i128) >= ((0) as i128))) && (false)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (!(((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) && (false)));
     assert!(!liquidate_otherwise(&mut s, i),
         "liquidate_otherwise must reject when guard is violated");
 }
